requirements.txt
httpx
psutil
websockets
pydantic
PyYAML
clickhouse-connect
